Add Nether Quartz
I am unsure if it was actually `quartz_ore` in the past, or if it was always `nether_quartz_ore`, but I have left the previous entry in, just in case. There's a possibility I just screwed up the name of quartz in the past, but never ran into the issue as I never really used Dog in the nether.